Feat(RESTful API): Created WasteLossRecordBatchFinishedProducts
WasteLossRecordBatchFinishedProducts has been implemented and is now functional. Initially I had an error with EF as I needed to name the constraint key it was trying to add as the character limit was exceeded.

I started by creating the model, add relationship in MystefyDBContext, create DTO, create interface, create the service then create the controller while referencing the DTO, Interface and service within the controller. Finally create a migration and update DataBase